backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
dj-database-url>=1.0
python-decouple>=3.8
psycopg2-binary>=2.9
gunicorn>=20.1
django-cors-headers>=3.14
python-dotenv


